requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-telegram-bot==21.9
anthropic==0.40.0
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.10.3
